radio: support 62.5 kHz LoRa bandwidth
The repeater only accepted 125/250/500 kHz. 62.5 kHz was blocked at
three layers: the config parser used integer-only parsing plus a
{125,250,500} whitelist, bandwidth/bw_khz were uint16_t (can't hold
the .5), and bw_code() had no 0x03 entry (unknown values silently
fell back to 250).

Make bandwidth a double (like frequency already is), add
LORA_BW_062 = 0x03, match in Hz so 62.5 resolves exactly, accept
{62.5,125,250,500} in the parser, and print with %g. Adds a
regression test that accepts 62.5 and rejects 62.

+++ src/config.h
@@ -17,7 +17,7 @@
 typedef struct {
     /* ---- radio ---- */
     double   frequency;          /* MHz */
-    uint16_t bandwidth;          /* kHz (125/250/500) */
+    double   bandwidth;          /* kHz (62.5/125/250/500) */
     uint8_t  spreading_factor;   /* 5..12 */
     uint8_t  coding_rate;        /* 5..8 (=4/5..4/8) */
     int8_t   tx_power;           /* SX1268 chip dBm (-9..22) */
